requirements.txt
amqp==5.2.0
asgiref==3.8.1
autopep8==2.1.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-extensions==3.2.3
django-silk==5.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gprof2dot==2022.7.29
idna==3.7
kombu==5.3.7
prompt-toolkit==3.0.43
pycodestyle==2.11.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
